PA-623

+++ projects/FieldNation/fieldnation/src/main/java/com/fieldnation/v2/data/model/ShipmentTask.java
@@ -25,7 +25,7 @@ public class ShipmentTask implements Parcelable {
     private static final String TAG = "ShipmentTask";
 
     @Json(name = "id")
-    private Double _id;
+    private Integer _id;
 
     @Source
     private JsonObject SOURCE;
@@ -38,15 +38,15 @@ public class ShipmentTask implements Parcelable {
         SOURCE = obj;
     }
 
-    public void setId(Double id) throws ParseException {
+    public void setId(Integer id) throws ParseException {
         _id = id;
         SOURCE.put("id", id);
     }
 
-    public Double getId() {
+    public Integer getId() {
         try {
             if (_id == null && SOURCE.has("id") && SOURCE.get("id") != null)
-                _id = SOURCE.getDouble("id");
+                _id = SOURCE.getInt("id");
         } catch (Exception ex) {
             Log.v(TAG, ex);
         }
@@ -54,7 +54,7 @@ public class ShipmentTask implements Parcelable {
         return _id;
     }
 
-    public ShipmentTask id(Double id) throws ParseException {
+    public ShipmentTask id(Integer id) throws ParseException {
         _id = id;
         SOURCE.put("id", id);
         return this;
